PLAYNEXT-1192 Fix responsibility between class and subclass to avoid exposing details to general purpose superclass
