+++ Application/Sources/RadioChannels/RadioChannelsViewController.swift
@@ -4,13 +4,19 @@
 //  License information is available from the LICENSE file.
 //
 
+import Combine
+import Tabman
 import UIKit
 
 @objc class RadioChannelsViewController: PageContainerViewController {
     private var radioChannelName: String?
+    private var satelliteRadioChannels: [RadioChannel] = []
+    private var cancellable: AnyCancellable?
 
     @objc init(radioChannels: [RadioChannel], satelliteRadioChannels: [RadioChannel]) {
         assert(!radioChannels.isEmpty, "At least 1 radio channel expected")
+
+        self.satelliteRadioChannels = satelliteRadioChannels
 
         var viewControllers = [UIViewController]()
         for (index, radioChannel) in radioChannels.enumerated() {
@@ -33,7 +39,7 @@ import UIKit
             NSNotFound
         }
 
-        super.init(viewControllers: viewControllers, additionalViewControllers: satelliteViewControllers, satelliteRadioChannels: satelliteRadioChannels, initialPage: initialPage)
+        super.init(viewControllers: viewControllers, additionalViewControllers: satelliteViewControllers, initialPage: initialPage)
         updateTitle()
     }
 
@@ -87,5 +93,28 @@ extension RadioChannelsViewController: PlayApplicationNavigation {
         }
 
         return false
+    }
+}
+
+// MARK: Swiss satellite radios
+
+extension RadioChannelsViewController {
+    private func srgMedia(for radioChannel: RadioChannel) -> AnyPublisher<SRGMedia, Error> {
+        SRGDataProvider.current!.regionalizedRadioLivestreams(for: ApplicationConfiguration.shared.vendor, contentProviders: .swissSatelliteRadio)
+            .compactMap { $0.first { $0.uid == radioChannel.uid } }
+            .eraseToAnyPublisher()
+    }
+
+    @objc override func tabDidChange(_ sender: TMTabItemBarButton) {
+        if sender.tag >= viewControllers.count {
+            cancellable = srgMedia(for: satelliteRadioChannels[sender.tag - viewControllers.count])
+                .receive(on: DispatchQueue.main)
+                .sink(
+                    receiveCompletion: { _ in },
+                    receiveValue: { [weak self] srgMedia in
+                        self?.play_presentMediaPlayer(with: srgMedia, position: nil, airPlaySuggestions: true, fromPushNotification: false, animated: true, completion: nil)
+                    }
+                )
+        }
     }
 }
